pyproject.toml
[project]
name = "pdf2md"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "backoff>=2.2.1",
    "openai>=1.75.0",
    "pdf2image>=1.17.0",
    "pillow>=11.2.1",
    "tqdm>=4.67.1",
]
